+++ ontask/dataops/services/canvas_upload.py
@@ -157,6 +157,37 @@ def _extract_quiz_answer_information(
                 data_frame_source)
 
 
+def _extract_enrollment_information(
+        oauth_info: dict,
+        user_token: models.OAuthUserToken,
+        canvas_course_id: int,
+        data_frame_source: dict):
+    """Extract enrollment information from Canvas and add to given data frame
+
+    :param oauth_info: Oauth info to connect to Canvas
+    :param user_token: User token to authorize the canvas connection
+    :param canvas_course_id: Canvas course
+    :param data_frame_source: Data frame to expand with the data
+    """
+
+    students = canvas_ops.request_and_access(
+        'get_course_enrolment',
+        oauth_info,
+        user_token,
+        [canvas_course_id, 'active'])
+
+    for student in students:
+        user_id = student['user']['id']
+        user_value = {
+            'id': user_id,
+            'name': student['user']['name']}
+        user_row = data_frame_source.get(user_id)
+        if not user_row:
+            user_row = user_value
+            data_frame_source[user_id] = user_row
+        user_row.update(user_value)
+
+
 def _extract_quiz_submission_information(
         oauth_info: dict,
         user_token: models.OAuthUserToken,
@@ -216,6 +247,10 @@ def _extract_assignment_submission_information(
         [canvas_course_id, assignment_id])
 
     for submission in assignment_submissions:
+        if not submission['attempt']:
+            # Submission with no attempts, skip
+            continue
+
         (
             submission_at_column_name,
             _,
@@ -236,10 +271,7 @@ def _extract_assignment_submission_information(
             data_frame_source[user_id] = user_row
 
         # Check if this is a more recent attempt
-        if (
-                submission['attempt'] and
-                submission['attempt'] > user_row.get(attempt_column_name, 0)
-        ):
+        if submission['attempt'] > user_row.get(attempt_column_name, 0):
             # This is a more recent submission, refresh data
             user_row[submission_at_column_name] = submission['submitted_at']
             user_row[submission_type] = submission['submission_type']
@@ -247,134 +279,102 @@ def _extract_assignment_submission_information(
             user_row[score_column] = submission['entered_score']
 
 
-def create_df_from_canvas_course_enrollment(
-        user: models.OnTaskUser,
-        target_url: str,
-        course_id: int
-) -> pd.DataFrame:
-    """Load data frame from a Canvas course enrollment list.
-
-    Given the Canvas target_url and a course ID, obtain the enrollment list
-    and load a dataframe with that list on the current workflow.
-
-    :param user: User object for authentication purposes.
-    :param target_url: Name of the Canvas instance to use.
-    :param course_id: Canvas Course ID (integer)
-    :return: Data Frame after obtaining it from Canvas.
-    """
-
-    # Verify parameter
-    canvas_course_id = canvas_ops.verify_course_id(course_id)
-
-    # Get the oauth info
-    oauth_info, user_token = canvas_ops.get_oauth_and_user_token(
-        user,
-        target_url)
-
-    students = canvas_ops.request_and_access(
-        'get_course_enrolment',
-        oauth_info,
-        user_token,
-        [canvas_course_id, 'active'])
-
-    data_frame_source = []
-    for student in students:
-        data_frame_source.append({
-            'id': student['user']['id'],
-            'canvas course id': course_id,
-            'name': student['user']['name']})
-
-    # Create the data frame with the collected data
-    return pd.DataFrame(data_frame_source)
-
-
-def create_df_from_canvas_course_quizzes(
+def create_df_from_canvas_course(
         user: models.OnTaskUser,
         target_url: str,
         course_id: int,
+        upload_enrollment: bool,
+        upload_quizzes: bool,
+        upload_assignments: bool,
+        include_course_id_column: bool,
         columns_to_upload: list = None
 ) -> pd.DataFrame:
-    """Load data frame from a Canvas with information about quizzes in a course.
+    """Create a data frame from the information in a canvas course
 
-    Given the Canvas target_url and a course ID, obtain all the information
-    related to quizzes in the course and create a pandas Data Frame.
-
-    :param user: User object for authentication purposes.
-    :param target_url: Name of the Canvas instance to use.
-    :param course_id: Canvas Course ID (integer)
-    :param columns_to_upload: List of columns to upload. Empty means upload all
-    :return: Data Frame after obtaining it from Canvas.
+    :param user: OnTaskUser performing the operation
+    :param target_url: URL of the Canvas server
+    :param course_id: Canvas course
+    :param upload_enrollment: Whether to upload the enrollment information
+    :param upload_quizzes: Whether to upload the quizzes information
+    :param upload_assignments: Whether to upload the assignments information
+    :param include_course_id_column: Whether to include a course id column
+    :param columns_to_upload: Columns in the workflow to upload
+    :return: The Pandas data frame with the extracted information
     """
     # Verify parameter
-    if columns_to_upload is None:
-        columns_to_upload = []
     canvas_course_id = canvas_ops.verify_course_id(course_id)
+
+    columns_to_upload = [] if columns_to_upload is None else columns_to_upload
 
     # Get the oauth info
     oauth_info, user_token = canvas_ops.get_oauth_and_user_token(
         user,
         target_url)
 
-    # Get list of active students enrolled in the course
-    students = canvas_ops.request_and_access(
-        'get_course_enrolment',
-        oauth_info,
-        user_token,
-        [course_id, 'active'])
-
+    # List of dictionaries to then create the Data Frame
     data_frame_source = {}
-    for student in students:
-        student_id = student['user']['id']
-        data_frame_source[student_id] = {
-            'id': student_id,
-            'canvas course id': course_id,
-            'name': student['user']['name']}
 
-    # Fetch all quizzes for the course
-    quizzes = canvas_ops.request_and_access(
-        'get_course_quizzes',
-        oauth_info,
-        user_token,
-        [course_id])
-
-    # Build the data frame source with the information from quizzes
-    for quiz in quizzes:
-        quiz_id = quiz['id']
-
-        # Get the answer information
-        _extract_quiz_answer_information(
+    # Upload Enrollments
+    if upload_enrollment:
+        _extract_enrollment_information(
             oauth_info,
             user_token,
             canvas_course_id,
-            quiz_id,
             data_frame_source)
 
-        # Get the quiz submission information
-        _extract_quiz_submission_information(
+    # Upload Quizzes
+    if upload_quizzes:
+        # Fetch all quizzes for the course
+        quizzes = canvas_ops.request_and_access(
+            'get_course_quizzes',
             oauth_info,
             user_token,
-            canvas_course_id,
-            quiz_id,
-            data_frame_source)
+            [course_id])
 
-    # Get all the assignments in a course
-    assignments = canvas_ops.request_and_access(
-        'get_course_assignments',
-        oauth_info,
-        user_token,
-        [canvas_course_id])
+        # Build the data frame source with the information from quizzes
+        for quiz in quizzes:
+            quiz_id = quiz['id']
 
-    for assignment in assignments:
-        # Get the assignment submission information
-        _extract_assignment_submission_information(
+            # Get the answer information
+            _extract_quiz_answer_information(
+                oauth_info,
+                user_token,
+                canvas_course_id,
+                quiz_id,
+                data_frame_source)
+
+            # Get the quiz submission information
+            _extract_quiz_submission_information(
+                oauth_info,
+                user_token,
+                canvas_course_id,
+                quiz_id,
+                data_frame_source)
+
+    # Upload Assignments
+    if upload_assignments:
+        # Get all the assignments in a course
+        assignments = canvas_ops.request_and_access(
+            'get_course_assignments',
             oauth_info,
             user_token,
-            canvas_course_id,
-            assignment['id'],
-            data_frame_source)
+            [canvas_course_id])
+
+        for assignment in assignments:
+            # Get the assignment submission information
+            _extract_assignment_submission_information(
+                oauth_info,
+                user_token,
+                canvas_course_id,
+                assignment['id'],
+                data_frame_source)
 
     # Create the data frame with the collected data
     result = pd.DataFrame(data_frame_source.values())
+
+    if include_course_id_column:
+        # Insert column with course id if required
+        result['canvas course id'] = canvas_course_id
 
     # Reorder the columns
     column_names = [
@@ -386,21 +386,58 @@ def create_df_from_canvas_course_quizzes(
         column_names = [
             cname for cname in column_names if cname in columns_to_upload]
 
+    head_column_names = ['id']
+    if upload_enrollment:
+        # If the enrollment has been updated, include name as head column
+        head_column_names.append('name')
+
+    if include_course_id_column:
+        # If the canvas course id column was added, include it as head column
+        head_column_names.append('canvas course id')
+
     # Sort the columns leaving ID as the first one
-    result = result[['id', 'canvas course id', 'name'] + sorted(column_names)]
+    result = result[head_column_names + sorted(column_names)]
 
     return result
 
 
-class ExecuteCanvasUploadBasic:
+class ExecuteCanvasCourseUpload:
     @staticmethod
-    def execute_dataframe_update(
-            user: models.OnTaskUser,
-            workflow: models.Workflow,
-            data_frame: pd.DataFrame,
-            payload: dict,
-            log_item: models.Log
+    def execute_operation(
+        user,
+        workflow: Optional[models.Workflow] = None,
+        action: Optional[models.Action] = None,
+        payload: Optional[Dict] = None,
+        log_item: Optional[models.Log] = None,
     ):
+        """Perform a Canvas Course Upload asynchronously.
+
+        :param user: User object
+        :param workflow: Workflow object
+        :param action: Empty
+        :param payload: has fields:
+          - dst_key: Key column in the existing dataframe (if any) for merge
+          - src_key: Key column in the external dataframe (for merge)
+          - how_merge: Merge method: inner, outer, left, right
+          - canvas_course_id: Unique course id in canvas
+          - target_url: URL for the remote canvas instance
+          - upload_enrollment: Whether to upload the enrollment information
+          - upload_quizzes: Whether to upload the quizzes information
+          - upload_assignments: Whether to upload the assignments information
+          - include_course_id_column: Whether to include a course id column
+          - columns_to_upload: Columns in the workflow to upload
+        :param log_item: Optional log item object.
+        """
+        data_frame = create_df_from_canvas_course(
+            user,
+            payload.pop('target_url'),
+            payload.pop('canvas_course_id'),
+            payload.pop('upload_enrollment'),
+            payload.pop('upload_quizzes'),
+            payload.pop('upload_assignments'),
+            payload.pop('include_course_id_column'),
+            payload.pop('columns_to_upload'))
+
         # Create session
         session = common.create_session()
 
@@ -414,99 +451,4 @@ class ExecuteCanvasUploadBasic:
             common.get_how_merge(payload, log_item),
             common.get_key(payload, 'src_key', log_item),
             common.get_key(payload, 'dst_key', log_item),
-            log_item)
-
-
-class ExecuteCanvasCourseEnrollmentsUpload(ExecuteCanvasUploadBasic):
-    """Process the Canvas Course Enrollment Upload operation in a workflow."""
-    def __init__(self):
-        """Assign default fields."""
-        super().__init__()
-        self.log_event = (
-            models.Log.WORKFLOW_DATA_CANVAS_COURSE_ENROLLMENT_UPLOAD)
-
-    @staticmethod
-    def execute_operation(
-        user,
-        workflow: Optional[models.Workflow] = None,
-        action: Optional[models.Action] = None,
-        payload: Optional[Dict] = None,
-        log_item: Optional[models.Log] = None,
-    ):
-        """Perform a Canvas Course Enrollment upload asynchronously.
-
-        :param user: User object
-        :param workflow: Workflow object
-        :param action: Empty
-        :param payload: has fields:
-          - dst_key: Key column in the existing dataframe (if any) for merge
-          - src_key: Key column in the external dataframe (for merge)
-          - how_merge: Merge method: inner, outer, left, right
-          - canvas_course_id: Unique course id in canvas
-          - target_url: URL for the remote canvas instance
-        :param log_item: Optional log item object.
-        """
-        # Create the data frame with the collected data
-        src_df = create_df_from_canvas_course_enrollment(
-            user,
-            payload.get('target_url'),
-            payload.get('canvas_course_id'))
-
-        # Update the dataframe
-        super(
-            ExecuteCanvasCourseEnrollmentsUpload,
-            ExecuteCanvasCourseEnrollmentsUpload
-        ).execute_dataframe_update(
-            user,
-            workflow,
-            src_df,
-            payload,
-            log_item)
-
-
-class ExecuteCanvasCourseQuizzesUpload(ExecuteCanvasUploadBasic):
-    """Process the Canvas Course Quizzes upload operation in a workflow."""
-
-    def __init__(self):
-        """Assign default fields."""
-        super().__init__()
-        self.log_event = models.Log.WORKFLOW_DATA_CANVAS_COURSE_QUIZZES_UPLOAD
-
-    @staticmethod
-    def execute_operation(
-        user,
-        workflow: Optional[models.Workflow] = None,
-        action: Optional[models.Action] = None,
-        payload: Optional[Dict] = None,
-        log_item: Optional[models.Log] = None,
-    ):
-        """Perform a Canvas upload asynchronously.
-
-        :param user: User object
-        :param workflow: Workflow object
-        :param action: Empty
-        :param payload: has fields:
-          - dst_key: Key column in the existing dataframe (if any) for merge
-          - src_key: Key column in the external dataframe (for merge)
-          - how_merge: Merge method: inner, outer, left, right
-          - canvas_course_id: Unique course id in canvas
-          - target_url: URL for the remote canvas instance
-        :param log_item: Optional log item object.
-        """
-        # Create the data frame with the collected data
-        src_df = create_df_from_canvas_course_quizzes(
-            user,
-            payload.get('target_url'),
-            payload.get('canvas_course_id'),
-            payload.get('columns_to_upload'))
-
-        # Update the dataframe
-        super(
-            ExecuteCanvasCourseQuizzesUpload,
-            ExecuteCanvasCourseQuizzesUpload
-        ).execute_dataframe_update(
-            user,
-            workflow,
-            src_df,
-            payload,
             log_item)