+++ ontask/dataops/views/canvas_upload.py
@@ -1,4 +1,4 @@
-"""First step for Canvas Course Enrollment upload."""
+"""First step for Canvas Course Upload."""
 from typing import Optional
 
 from django import http
@@ -20,7 +20,6 @@ class CanvasUploadStart(upload_steps.UploadStepOneView):
     The different operations are defined by parameters used when invoking
     this as a view. See url.py in this module.
     """
-
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
         # Needed for authentication purposes
@@ -33,9 +32,13 @@ class CanvasUploadStart(upload_steps.UploadStepOneView):
             self.request,
             {
                 'step_1': reverse(self.step_1_url),
-                'log_upload': self.log_upload,
                 'target_url': form.cleaned_data['target_url'],
-                'canvas_course_id': form.cleaned_data['canvas_course_id']})
+                'canvas_course_id': form.cleaned_data['canvas_course_id'],
+                'upload_enrollment': form.cleaned_data['upload_enrollment'],
+                'upload_quizzes': form.cleaned_data['upload_quizzes'],
+                'upload_assignments': form.cleaned_data['upload_assignments'],
+                'include_course_id_column': form.cleaned_data[
+                    'include_course_id_column']})
 
         # Go fetch the OAuth token, callback is canvas_upload_start_finish
         return canvas_ops.set_oauth_token(
@@ -60,24 +63,14 @@ def canvas_upload_start_finish(
     # Process Canvas API call to get the list of students
     payload = session_ops.get_payload(request)
     try:
-        op_type = payload.pop('log_upload')
-        target_url = payload.pop('target_url')
-        course_id = payload.pop('canvas_course_id')
-        if (op_type ==
-                models.Log.WORKFLOW_DATA_CANVAS_COURSE_ENROLLMENT_UPLOAD):
-            data_frame = services.create_df_from_canvas_course_enrollment(
-                    request.user,
-                    target_url,
-                    course_id)
-        elif (op_type ==
-              models.Log.WORKFLOW_DATA_CANVAS_COURSE_QUIZZES_UPLOAD):
-            data_frame = services.create_df_from_canvas_course_quizzes(
-                request.user,
-                target_url,
-                course_id)
-        else:
-            raise OnTaskException(
-                _('Unexpected Canvas operation not supported.'))
+        data_frame = services.create_df_from_canvas_course(
+            request.user,
+            payload.pop('target_url'),
+            payload.pop('canvas_course_id'),
+            payload.pop('upload_enrollment'),
+            payload.pop('upload_quizzes'),
+            payload.pop('upload_assignments'),
+            payload.pop('include_course_id_column'))
 
         if data_frame.empty:
             raise OnTaskException(
@@ -89,7 +82,7 @@ def canvas_upload_start_finish(
     except Exception as exc:
         messages.error(
             request,
-            _('Canvas course upload was not successful: {0}').format(
+            _('Canvas upload was not successful: {0}').format(
                 str(exc)))
         return redirect('dataops:uploadmerge')
 
@@ -101,8 +94,8 @@ def canvas_upload_start_finish(
             'initial_column_names': frame_info[0],
             'column_types': frame_info[1],
             'src_is_key_column': frame_info[2],
-            'step_1': reverse('dataops:canvas_course_enrollments_upload_start'),
+            'step_1': reverse('dataops:canvas_course_upload_start'),
             'log_upload':
-                models.Log.WORKFLOW_DATA_CANVAS_COURSE_ENROLLMENT_UPLOAD})
+                models.Log.WORKFLOW_DATA_CANVAS_COURSE_UPLOAD})
 
     return redirect('dataops:upload_s2')