+++ ontask/scheduler/forms/__init__.py
@@ -1,7 +1,7 @@
 """Import the forms"""
 from ontask.scheduler.forms.basic import (
     ScheduleBasicForm, ScheduleBasicUpdateForm, ScheduleSQLUploadForm,
-    ScheduleUploadCanvasForm, ScheduleUploadCanvasQuizForm)
+    ScheduleUploadCanvasForm)
 from ontask.scheduler.forms.run import (
     ScheduleEmailForm, ScheduleJSONForm, ScheduleJSONReportForm,
     ScheduleSendListForm, ScheduleCanvasEmailForm