+++ ontask/scheduler/services/__init__.py
@@ -11,6 +11,5 @@ from ontask.scheduler.services.sql_producer import (
     ScheduledOperationUpdateSQLUpload,
 )
 from ontask.scheduler.services.canvas_update_producer import (
-    ScheduledOperationUpdateCanvasCourseQuizzesUpload,
-    ScheduledOperationUpdateCanvasCourseEnrollmentUpload)
+    ScheduledOperationUpdateCanvasUpload)
 from ontask.scheduler.services.tasks_ops import schedule_task